+++ rootfs/etc/owncloud.d/45-php.sh
@@ -1,5 +1,10 @@
 #!/usr/bin/env bash
 
+echo "Writing objectstore config..."
+gomplate \
+  -f /etc/templates/objectstore.php \
+  -o ${OWNCLOUD_VOLUME_CONFIG}/objectstore.config.php
+
 echo "Writing php config..."
 gomplate \
   -f /etc/templates/owncloud.ini \